requirements.txt
streamlit
yfinance
google-generativeai
gtts